final commit

+++ KomodoCafe_ClassLibrary/Menu.cs
@@ -13,14 +13,14 @@ namespace KomodoCafe_ClassLibrary
         public int MealNumber { get; set; }
         public string MealName { get; set; }
         public string MealDescription { get; set; }
-        public decimal MealPrice { get; set; }
+        public double MealPrice { get; set; }
         public string MealIngredients { get; set; }
 
 
 
         public Menu() { }
 
-        public Menu(int mealNumber, string mealName, string mealDescription, decimal mealPrice, string mealIngredients)
+        public Menu(int mealNumber, string mealName, string mealDescription, double mealPrice, string mealIngredients)
         {
             MealNumber = mealNumber;
             MealName = mealName;
